fraud-detection-system_copy/backend/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
bcrypt
PyJWT